+++ tests/client_test.rs
@@ -3,10 +3,10 @@ use embedded_io::BufRead as _;
 use embedded_io_adapters::{std::FromStd, tokio_1::FromTokio};
 use embedded_io_async::BufRead as _;
 use embedded_io_async::Write;
-use rand::rngs::OsRng;
 use std::net::SocketAddr;
 use std::sync::Once;
 
+mod common;
 mod tlsserver;
 
 static LOG_INIT: Once = Once::new();
@@ -57,16 +57,13 @@ async fn test_google() {
     let mut write_record_buffer = [0; 16384];
     let config = TlsConfig::new().with_server_name("google.com");
 
-    let mut tls = TlsConnection::new(
+    let mut tls: TlsConnection<_, Aes128GcmSha256> = TlsConnection::new(
         FromTokio::new(stream),
         &mut read_record_buffer,
         &mut write_record_buffer,
     );
 
-    let open_fut = tls.open(TlsContext::new(
-        &config,
-        UnsecureProvider::new::<Aes128GcmSha256>(OsRng),
-    ));
+    let open_fut = tls.open(TlsContext::new(&config));
     log::info!("SIZE of open fut is {}", core::mem::size_of_val(&open_fut));
     open_fut.await.expect("error establishing TLS connection");
     log::info!("Established");
@@ -101,7 +98,7 @@ async fn test_ping() {
     let mut write_record_buffer = [0; 16384];
     let config = TlsConfig::new().with_server_name("localhost");
 
-    let mut tls = TlsConnection::new(
+    let mut tls: TlsConnection<_, Aes128GcmSha256> = TlsConnection::new(
         FromTokio::new(stream),
         &mut read_record_buffer,
         &mut write_record_buffer,
@@ -109,10 +106,7 @@ async fn test_ping() {
 
     log::info!("SIZE of connection is {}", core::mem::size_of_val(&tls));
 
-    let open_fut = tls.open(TlsContext::new(
-        &config,
-        UnsecureProvider::new::<Aes128GcmSha256>(OsRng),
-    ));
+    let open_fut = tls.open(TlsContext::new(&config));
     log::info!("SIZE of open fut is {}", core::mem::size_of_val(&open_fut));
     open_fut.await.expect("error establishing TLS connection");
     log::info!("Established");
@@ -166,7 +160,7 @@ async fn test_ping_nocopy() {
     let mut write_record_buffer = [0; 16384];
     let config = TlsConfig::new().with_server_name("localhost");
 
-    let mut tls = TlsConnection::new(
+    let mut tls: TlsConnection<_, Aes128GcmSha256> = TlsConnection::new(
         FromTokio::new(stream),
         &mut read_record_buffer,
         &mut write_record_buffer,
@@ -174,10 +168,7 @@ async fn test_ping_nocopy() {
 
     log::info!("SIZE of connection is {}", core::mem::size_of_val(&tls));
 
-    let open_fut = tls.open(TlsContext::new(
-        &config,
-        UnsecureProvider::new::<Aes128GcmSha256>(OsRng),
-    ));
+    let open_fut = tls.open(TlsContext::new(&config));
     log::info!("SIZE of open fut is {}", core::mem::size_of_val(&open_fut));
     open_fut.await.expect("error establishing TLS connection");
     log::info!("Established");
@@ -234,17 +225,14 @@ async fn test_ping_nocopy_bufread() {
     let mut write_record_buffer = [0; 16384];
     let config = TlsConfig::new().with_server_name("localhost");
 
-    let mut tls = TlsConnection::new(
+    let mut tls: TlsConnection<_, Aes128GcmSha256> = TlsConnection::new(
         FromTokio::new(stream),
         &mut read_record_buffer,
         &mut write_record_buffer,
     );
-    tls.open(TlsContext::new(
-        &config,
-        UnsecureProvider::new::<Aes128GcmSha256>(OsRng),
-    ))
-    .await
-    .expect("error establishing TLS connection");
+    tls.open(TlsContext::new(&config))
+        .await
+        .expect("error establishing TLS connection");
     log::info!("Established");
 
     tls.write(b"ping").await.expect("error writing data");
@@ -282,11 +270,8 @@ fn test_blocking_ping() {
         &mut read_record_buffer,
         &mut write_record_buffer,
     );
-    tls.open(TlsContext::new(
-        &config,
-        UnsecureProvider::new::<Aes128GcmSha256>(OsRng),
-    ))
-    .expect("error establishing TLS connection");
+    tls.open(TlsContext::new(&config))
+        .expect("error establishing TLS connection");
     log::info!("Established");
 
     tls.write(b"ping").expect("error writing data");
@@ -331,11 +316,8 @@ fn test_blocking_ping_nocopy() {
         &mut read_record_buffer,
         &mut write_record_buffer,
     );
-    tls.open(TlsContext::new(
-        &config,
-        UnsecureProvider::new::<Aes128GcmSha256>(OsRng),
-    ))
-    .expect("error establishing TLS connection");
+    tls.open(TlsContext::new(&config))
+        .expect("error establishing TLS connection");
     log::info!("Established");
 
     tls.write(b"ping").expect("error writing data");
@@ -374,11 +356,8 @@ fn test_blocking_ping_nocopy_bufread() {
         &mut read_record_buffer,
         &mut write_record_buffer,
     );
-    tls.open(TlsContext::new(
-        &config,
-        UnsecureProvider::new::<Aes128GcmSha256>(OsRng),
-    ))
-    .expect("error establishing TLS connection");
+    tls.open(TlsContext::new(&config))
+        .expect("error establishing TLS connection");
     log::info!("Established");
 
     tls.write(b"ping").expect("error writing data");